[AIP-21] Move operators.s3_to_redshift_operator providers.amazon.aws.operators.s3_to_redshift
